+++ changelog.md
@@ -1,4 +1,6 @@
 # v1.0.2
 
+ - add generated file to source set
  - fix tests
  - fix release ci
+ - add support to auto executing task on project build